LayoutTests/imported/w3c/web-platform-tests/html/syntax/serializing-html-fragments/serializing-expected.txt


PASS innerHTML 0  
PASS innerHTML 1 <a></a> 
PASS innerHTML 2 <a b="c"></a> 
PASS innerHTML 3 <a b="c"></a> 
PASS innerHTML 4 <a b="&amp;"></a> 
PASS innerHTML 5 <a b="&nbsp;"></a> 
PASS innerHTML 6 <a b="&quot;"></a> 
PASS innerHTML 7 <a b="<"></a> 
PASS innerHTML 8 <a b=">"></a> 
PASS innerHTML 9 <svg xlink:href="a"></svg> 
PASS innerHTML 10 <svg xmlns:svg="test"></svg> 
PASS innerHTML 11 a 
PASS innerHTML 12 &amp; 
PASS innerHTML 13 &nbsp; 
PASS innerHTML 14 &lt; 
PASS innerHTML 15 &gt; 
PASS innerHTML 16 " 
PASS innerHTML 17 <style><&></style> 
PASS innerHTML 18 <script type="test"><&></script> 
PASS innerHTML 19 <xmp><&></xmp> 
FAIL innerHTML 20 <iframe><&></iframe> assert_equals: expected "<iframe><&></iframe>" but got "<iframe>&lt;&amp;&gt;</iframe>"
FAIL innerHTML 21 <noembed><&></noembed> assert_equals: expected "<noembed><&></noembed>" but got "<noembed>&lt;&amp;&gt;</noembed>"
FAIL innerHTML 22 <noframes><&></noframes> assert_equals: expected "<noframes><&></noframes>" but got "<noframes>&lt;&amp;&gt;</noframes>"
FAIL innerHTML 23 <noscript><&></noscript> assert_equals: expected "<noscript><&></noscript>" but got "<noscript>&lt;&amp;&gt;</noscript>"
FAIL innerHTML 24 <!--data--> assert_equals: expected "<!--data-->" but got "<!-- data -->"
PASS innerHTML 25 <a><b><c></c></b><d>e</d><f><g>h</g></f></a> 
PASS innerHTML 26  
PASS outerHTML 0 <span></span> 
PASS outerHTML 1 <span><a></a></span> 
PASS outerHTML 2 <span><a b="c"></a></span> 
PASS outerHTML 3 <span><a b="c"></a></span> 
PASS outerHTML 4 <span><a b="&amp;"></a></span> 
PASS outerHTML 5 <span><a b="&nbsp;"></a></span> 
PASS outerHTML 6 <span><a b="&quot;"></a></span> 
PASS outerHTML 7 <span><a b="<"></a></span> 
PASS outerHTML 8 <span><a b=">"></a></span> 
PASS outerHTML 9 <span><svg xlink:href="a"></svg></span> 
PASS outerHTML 10 <span><svg xmlns:svg="test"></svg></span> 
PASS outerHTML 11 <span>a</span> 
PASS outerHTML 12 <span>&amp;</span> 
PASS outerHTML 13 <span>&nbsp;</span> 
PASS outerHTML 14 <span>&lt;</span> 
PASS outerHTML 15 <span>&gt;</span> 
PASS outerHTML 16 <span>"</span> 
PASS outerHTML 17 <span><style><&></style></span> 
PASS outerHTML 18 <span><script type="test"><&></script></span> 
PASS outerHTML 19 <span><xmp><&></xmp></span> 
FAIL outerHTML 20 <span><iframe><&></iframe></span> assert_equals: expected "<span><iframe><&></iframe></span>" but got "<span><iframe>&lt;&amp;&gt;</iframe></span>"
FAIL outerHTML 21 <span><noembed><&></noembed></span> assert_equals: expected "<span><noembed><&></noembed></span>" but got "<span><noembed>&lt;&amp;&gt;</noembed></span>"
FAIL outerHTML 22 <span><noframes><&></noframes></span> assert_equals: expected "<span><noframes><&></noframes></span>" but got "<span><noframes>&lt;&amp;&gt;</noframes></span>"
FAIL outerHTML 23 <span><noscript><&></noscript></span> assert_equals: expected "<span><noscript><&></noscript></span>" but got "<span><noscript>&lt;&amp;&gt;</noscript></span>"
FAIL outerHTML 24 <span><!--data--></span> assert_equals: expected "<span><!--data--></span>" but got "<span><!-- data --></span>"
PASS outerHTML 25 <span><a><b><c></c></b><d>e</d><f><g>h</g></f></a></span> 
PASS outerHTML 26 <span b="c"></span> 
PASS innerHTML Attribute in the XML namespace 
FAIL innerHTML Attribute in the XML namespace with the prefix not set to xml: assert_equals: expected "<svg xml:foo=\"test\"></svg>" but got "<svg abc:foo=\"test\"></svg>"
PASS innerHTML Non-'xmlns' attribute in the xmlns namespace 
PASS innerHTML 'xmlns' attribute in the xmlns namespace 
FAIL innerHTML Attribute in non-standard namespace assert_equals: expected "<svg abc:def=\"test\"></svg>" but got "<svg def=\"test\"></svg>"
PASS innerHTML <span> starting with U+000A 
PASS outerHTML Attribute in the XML namespace 
FAIL outerHTML Attribute in the XML namespace with the prefix not set to xml: assert_equals: expected "<span><svg xml:foo=\"test\"></svg></span>" but got "<span><svg abc:foo=\"test\"></svg></span>"
PASS outerHTML Non-'xmlns' attribute in the xmlns namespace 
PASS outerHTML 'xmlns' attribute in the xmlns namespace 
FAIL outerHTML Attribute in non-standard namespace assert_equals: expected "<span><svg abc:def=\"test\"></svg></span>" but got "<span><svg def=\"test\"></svg></span>"
PASS outerHTML <span> starting with U+000A 
PASS innerHTML <pre> context starting with U+000A 
PASS innerHTML <textarea> context starting with U+000A 
PASS innerHTML <listing> context starting with U+000A 
PASS innerHTML <pre> context not starting with U+000A 
PASS innerHTML <textarea> context not starting with U+000A 
PASS innerHTML <listing> context not starting with U+000A 
FAIL innerHTML <pre> non-context starting with U+000A assert_equals: expected "<pre>\n\n</%text>" but got "<pre>\n</pre>"
FAIL innerHTML <textarea> non-context starting with U+000A assert_equals: expected "<textarea>\n\n</%text>" but got "<textarea>\n</textarea>"
FAIL innerHTML <listing> non-context starting with U+000A assert_equals: expected "<listing>\n\n</%text>" but got "<listing>\n</listing>"
FAIL innerHTML <pre> non-context not starting with U+000A assert_equals: expected "<pre>a\n</%text>" but got "<pre>a\n</pre>"
FAIL innerHTML <textarea> non-context not starting with U+000A assert_equals: expected "<textarea>a\n</%text>" but got "<textarea>a\n</textarea>"
FAIL innerHTML <listing> non-context not starting with U+000A assert_equals: expected "<listing>a\n</%text>" but got "<listing>a\n</listing>"
FAIL outerHTML <pre> context starting with U+000A assert_equals: expected "<pre>\n\n</%text>" but got "<pre>\n</pre>"
FAIL outerHTML <textarea> context starting with U+000A assert_equals: expected "<textarea>\n\n</%text>" but got "<textarea>\n</textarea>"
FAIL outerHTML <listing> context starting with U+000A assert_equals: expected "<listing>\n\n</%text>" but got "<listing>\n</listing>"
FAIL outerHTML <pre> context not starting with U+000A assert_equals: expected "<pre>a\n</%text>" but got "<pre>a\n</pre>"
FAIL outerHTML <textarea> context not starting with U+000A assert_equals: expected "<textarea>a\n</%text>" but got "<textarea>a\n</textarea>"
FAIL outerHTML <listing> context not starting with U+000A assert_equals: expected "<listing>a\n</%text>" but got "<listing>a\n</listing>"
FAIL outerHTML <pre> non-context starting with U+000A assert_equals: expected "<span><pre>\n\n</%text></span>" but got "<span><pre>\n</pre></span>"
FAIL outerHTML <textarea> non-context starting with U+000A assert_equals: expected "<span><textarea>\n\n</%text></span>" but got "<span><textarea>\n</textarea></span>"
FAIL outerHTML <listing> non-context starting with U+000A assert_equals: expected "<span><listing>\n\n</%text></span>" but got "<span><listing>\n</listing></span>"
FAIL outerHTML <pre> non-context not starting with U+000A assert_equals: expected "<span><pre>a\n</%text></span>" but got "<span><pre>a\n</pre></span>"
FAIL outerHTML <textarea> non-context not starting with U+000A assert_equals: expected "<span><textarea>a\n</%text></span>" but got "<span><textarea>a\n</textarea></span>"
FAIL outerHTML <listing> non-context not starting with U+000A assert_equals: expected "<span><listing>a\n</%text></span>" but got "<span><listing>a\n</listing></span>"
PASS innerHTML Void context node area 
PASS innerHTML Void context node base 
PASS innerHTML Void context node basefont 
PASS innerHTML Void context node bgsound 
PASS innerHTML Void context node br 
PASS innerHTML Void context node col 
PASS innerHTML Void context node embed 
PASS innerHTML Void context node frame 
PASS innerHTML Void context node hr 
PASS innerHTML Void context node img 
PASS innerHTML Void context node input 
PASS innerHTML Void context node keygen 
PASS innerHTML Void context node link 
FAIL innerHTML Void context node menuitem assert_equals: expected "" but got "<a><!--abc--></a><b><c>abc</c></b>"
PASS innerHTML Void context node meta 
PASS innerHTML Void context node param 
PASS innerHTML Void context node source 
PASS innerHTML Void context node track 
PASS innerHTML Void context node wbr 
FAIL innerHTML void as first child with following siblings area assert_equals: expected "<area><a>test</a><b></b>" but got "<area></area><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings base assert_equals: expected "<base><a>test</a><b></b>" but got "<base></base><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings basefont assert_equals: expected "<basefont><a>test</a><b></b>" but got "<basefont></basefont><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings bgsound assert_equals: expected "<bgsound><a>test</a><b></b>" but got "<bgsound></bgsound><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings br assert_equals: expected "<br><a>test</a><b></b>" but got "<br></br><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings col assert_equals: expected "<col><a>test</a><b></b>" but got "<col></col><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings embed assert_equals: expected "<embed><a>test</a><b></b>" but got "<embed></embed><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings frame assert_equals: expected "<frame><a>test</a><b></b>" but got "<frame></frame><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings hr assert_equals: expected "<hr><a>test</a><b></b>" but got "<hr></hr><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings img assert_equals: expected "<img><a>test</a><b></b>" but got "<img></img><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings input assert_equals: expected "<input><a>test</a><b></b>" but got "<input></input><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings keygen assert_equals: expected "<keygen><a>test</a><b></b>" but got "<keygen></keygen><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings link assert_equals: expected "<link><a>test</a><b></b>" but got "<link></link><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings menuitem assert_equals: expected "<menuitem><a>test</a><b></b>" but got "<menuitem><a><!--abc--></a><b><c>abc</c></b></menuitem><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings meta assert_equals: expected "<meta><a>test</a><b></b>" but got "<meta></meta><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings param assert_equals: expected "<param><a>test</a><b></b>" but got "<param></param><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings source assert_equals: expected "<source><a>test</a><b></b>" but got "<source></source><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings track assert_equals: expected "<track><a>test</a><b></b>" but got "<track></track><a>test</a><b></b>"
FAIL innerHTML void as first child with following siblings wbr assert_equals: expected "<wbr><a>test</a><b></b>" but got "<wbr></wbr><a>test</a><b></b>"
FAIL innerHTML void as second child with following siblings area assert_equals: expected "<a>test</a><area><b></b>" but got "<a>test</a><area></area><b></b>"
FAIL innerHTML void as second child with following siblings base assert_equals: expected "<a>test</a><base><b></b>" but got "<a>test</a><base></base><b></b>"
FAIL innerHTML void as second child with following siblings basefont assert_equals: expected "<a>test</a><basefont><b></b>" but got "<a>test</a><basefont></basefont><b></b>"
FAIL innerHTML void as second child with following siblings bgsound assert_equals: expected "<a>test</a><bgsound><b></b>" but got "<a>test</a><bgsound></bgsound><b></b>"
FAIL innerHTML void as second child with following siblings br assert_equals: expected "<a>test</a><br><b></b>" but got "<a>test</a><br></br><b></b>"
FAIL innerHTML void as second child with following siblings col assert_equals: expected "<a>test</a><col><b></b>" but got "<a>test</a><col></col><b></b>"
FAIL innerHTML void as second child with following siblings embed assert_equals: expected "<a>test</a><embed><b></b>" but got "<a>test</a><embed></embed><b></b>"
FAIL innerHTML void as second child with following siblings frame assert_equals: expected "<a>test</a><frame><b></b>" but got "<a>test</a><frame></frame><b></b>"
FAIL innerHTML void as second child with following siblings hr assert_equals: expected "<a>test</a><hr><b></b>" but got "<a>test</a><hr></hr><b></b>"
FAIL innerHTML void as second child with following siblings img assert_equals: expected "<a>test</a><img><b></b>" but got "<a>test</a><img></img><b></b>"
FAIL innerHTML void as second child with following siblings input assert_equals: expected "<a>test</a><input><b></b>" but got "<a>test</a><input></input><b></b>"
FAIL innerHTML void as second child with following siblings keygen assert_equals: expected "<a>test</a><keygen><b></b>" but got "<a>test</a><keygen></keygen><b></b>"
FAIL innerHTML void as second child with following siblings link assert_equals: expected "<a>test</a><link><b></b>" but got "<a>test</a><link></link><b></b>"
FAIL innerHTML void as second child with following siblings menuitem assert_equals: expected "<a>test</a><menuitem><b></b>" but got "<a>test</a><menuitem><a><!--abc--></a><b><c>abc</c></b></menuitem><b></b>"
FAIL innerHTML void as second child with following siblings meta assert_equals: expected "<a>test</a><meta><b></b>" but got "<a>test</a><meta></meta><b></b>"
FAIL innerHTML void as second child with following siblings param assert_equals: expected "<a>test</a><param><b></b>" but got "<a>test</a><param></param><b></b>"
FAIL innerHTML void as second child with following siblings source assert_equals: expected "<a>test</a><source><b></b>" but got "<a>test</a><source></source><b></b>"
FAIL innerHTML void as second child with following siblings track assert_equals: expected "<a>test</a><track><b></b>" but got "<a>test</a><track></track><b></b>"
FAIL innerHTML void as second child with following siblings wbr assert_equals: expected "<a>test</a><wbr><b></b>" but got "<a>test</a><wbr></wbr><b></b>"
FAIL innerHTML void as last child with preceding siblings area assert_equals: expected "<a>test</a><b></b><area>" but got "<a>test</a><b></b><area></area>"
FAIL innerHTML void as last child with preceding siblings base assert_equals: expected "<a>test</a><b></b><base>" but got "<a>test</a><b></b><base></base>"
FAIL innerHTML void as last child with preceding siblings basefont assert_equals: expected "<a>test</a><b></b><basefont>" but got "<a>test</a><b></b><basefont></basefont>"
FAIL innerHTML void as last child with preceding siblings bgsound assert_equals: expected "<a>test</a><b></b><bgsound>" but got "<a>test</a><b></b><bgsound></bgsound>"
FAIL innerHTML void as last child with preceding siblings br assert_equals: expected "<a>test</a><b></b><br>" but got "<a>test</a><b></b><br></br>"
FAIL innerHTML void as last child with preceding siblings col assert_equals: expected "<a>test</a><b></b><col>" but got "<a>test</a><b></b><col></col>"
FAIL innerHTML void as last child with preceding siblings embed assert_equals: expected "<a>test</a><b></b><embed>" but got "<a>test</a><b></b><embed></embed>"
FAIL innerHTML void as last child with preceding siblings frame assert_equals: expected "<a>test</a><b></b><frame>" but got "<a>test</a><b></b><frame></frame>"
FAIL innerHTML void as last child with preceding siblings hr assert_equals: expected "<a>test</a><b></b><hr>" but got "<a>test</a><b></b><hr></hr>"
FAIL innerHTML void as last child with preceding siblings img assert_equals: expected "<a>test</a><b></b><img>" but got "<a>test</a><b></b><img></img>"
FAIL innerHTML void as last child with preceding siblings input assert_equals: expected "<a>test</a><b></b><input>" but got "<a>test</a><b></b><input></input>"
FAIL innerHTML void as last child with preceding siblings keygen assert_equals: expected "<a>test</a><b></b><keygen>" but got "<a>test</a><b></b><keygen></keygen>"
FAIL innerHTML void as last child with preceding siblings link assert_equals: expected "<a>test</a><b></b><link>" but got "<a>test</a><b></b><link></link>"
FAIL innerHTML void as last child with preceding siblings menuitem assert_equals: expected "<a>test</a><b></b><menuitem>" but got "<a>test</a><b></b><menuitem><a><!--abc--></a><b><c>abc</c></b></menuitem>"
FAIL innerHTML void as last child with preceding siblings meta assert_equals: expected "<a>test</a><b></b><meta>" but got "<a>test</a><b></b><meta></meta>"
FAIL innerHTML void as last child with preceding siblings param assert_equals: expected "<a>test</a><b></b><param>" but got "<a>test</a><b></b><param></param>"
FAIL innerHTML void as last child with preceding siblings source assert_equals: expected "<a>test</a><b></b><source>" but got "<a>test</a><b></b><source></source>"
FAIL innerHTML void as last child with preceding siblings track assert_equals: expected "<a>test</a><b></b><track>" but got "<a>test</a><b></b><track></track>"
FAIL innerHTML void as last child with preceding siblings wbr assert_equals: expected "<a>test</a><b></b><wbr>" but got "<a>test</a><b></b><wbr></wbr>"
FAIL outerHTML Void context node area assert_equals: expected "<area>" but got "<area></area>"
FAIL outerHTML Void context node base assert_equals: expected "<base>" but got "<base></base>"
FAIL outerHTML Void context node basefont assert_equals: expected "<basefont>" but got "<basefont></basefont>"
FAIL outerHTML Void context node bgsound assert_equals: expected "<bgsound>" but got "<bgsound></bgsound>"
FAIL outerHTML Void context node br assert_equals: expected "<br>" but got "<br></br>"
FAIL outerHTML Void context node col assert_equals: expected "<col>" but got "<col></col>"
FAIL outerHTML Void context node embed assert_equals: expected "<embed>" but got "<embed></embed>"
FAIL outerHTML Void context node frame assert_equals: expected "<frame>" but got "<frame></frame>"
FAIL outerHTML Void context node hr assert_equals: expected "<hr>" but got "<hr></hr>"
FAIL outerHTML Void context node img assert_equals: expected "<img>" but got "<img></img>"
FAIL outerHTML Void context node input assert_equals: expected "<input>" but got "<input></input>"
FAIL outerHTML Void context node keygen assert_equals: expected "<keygen>" but got "<keygen></keygen>"
FAIL outerHTML Void context node link assert_equals: expected "<link>" but got "<link></link>"
FAIL outerHTML Void context node menuitem assert_equals: expected "<menuitem>" but got "<menuitem><a><!--abc--></a><b><c>abc</c></b></menuitem>"
FAIL outerHTML Void context node meta assert_equals: expected "<meta>" but got "<meta></meta>"
FAIL outerHTML Void context node param assert_equals: expected "<param>" but got "<param></param>"
FAIL outerHTML Void context node source assert_equals: expected "<source>" but got "<source></source>"
FAIL outerHTML Void context node track assert_equals: expected "<track>" but got "<track></track>"
FAIL outerHTML Void context node wbr assert_equals: expected "<wbr>" but got "<wbr></wbr>"
FAIL outerHTML void as first child with following siblings area assert_equals: expected "<span><area><a>test</a><b></b></span>" but got "<span><area></area><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings base assert_equals: expected "<span><base><a>test</a><b></b></span>" but got "<span><base></base><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings basefont assert_equals: expected "<span><basefont><a>test</a><b></b></span>" but got "<span><basefont></basefont><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings bgsound assert_equals: expected "<span><bgsound><a>test</a><b></b></span>" but got "<span><bgsound></bgsound><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings br assert_equals: expected "<span><br><a>test</a><b></b></span>" but got "<span><br></br><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings col assert_equals: expected "<span><col><a>test</a><b></b></span>" but got "<span><col></col><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings embed assert_equals: expected "<span><embed><a>test</a><b></b></span>" but got "<span><embed></embed><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings frame assert_equals: expected "<span><frame><a>test</a><b></b></span>" but got "<span><frame></frame><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings hr assert_equals: expected "<span><hr><a>test</a><b></b></span>" but got "<span><hr></hr><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings img assert_equals: expected "<span><img><a>test</a><b></b></span>" but got "<span><img></img><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings input assert_equals: expected "<span><input><a>test</a><b></b></span>" but got "<span><input></input><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings keygen assert_equals: expected "<span><keygen><a>test</a><b></b></span>" but got "<span><keygen></keygen><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings link assert_equals: expected "<span><link><a>test</a><b></b></span>" but got "<span><link></link><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings menuitem assert_equals: expected "<span><menuitem><a>test</a><b></b></span>" but got "<span><menuitem><a><!--abc--></a><b><c>abc</c></b></menuitem><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings meta assert_equals: expected "<span><meta><a>test</a><b></b></span>" but got "<span><meta></meta><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings param assert_equals: expected "<span><param><a>test</a><b></b></span>" but got "<span><param></param><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings source assert_equals: expected "<span><source><a>test</a><b></b></span>" but got "<span><source></source><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings track assert_equals: expected "<span><track><a>test</a><b></b></span>" but got "<span><track></track><a>test</a><b></b></span>"
FAIL outerHTML void as first child with following siblings wbr assert_equals: expected "<span><wbr><a>test</a><b></b></span>" but got "<span><wbr></wbr><a>test</a><b></b></span>"
FAIL outerHTML void as second child with following siblings area assert_equals: expected "<span><a>test</a><area><b></b></span>" but got "<span><a>test</a><area></area><b></b></span>"
FAIL outerHTML void as second child with following siblings base assert_equals: expected "<span><a>test</a><base><b></b></span>" but got "<span><a>test</a><base></base><b></b></span>"
FAIL outerHTML void as second child with following siblings basefont assert_equals: expected "<span><a>test</a><basefont><b></b></span>" but got "<span><a>test</a><basefont></basefont><b></b></span>"
FAIL outerHTML void as second child with following siblings bgsound assert_equals: expected "<span><a>test</a><bgsound><b></b></span>" but got "<span><a>test</a><bgsound></bgsound><b></b></span>"
FAIL outerHTML void as second child with following siblings br assert_equals: expected "<span><a>test</a><br><b></b></span>" but got "<span><a>test</a><br></br><b></b></span>"
FAIL outerHTML void as second child with following siblings col assert_equals: expected "<span><a>test</a><col><b></b></span>" but got "<span><a>test</a><col></col><b></b></span>"
FAIL outerHTML void as second child with following siblings embed assert_equals: expected "<span><a>test</a><embed><b></b></span>" but got "<span><a>test</a><embed></embed><b></b></span>"
FAIL outerHTML void as second child with following siblings frame assert_equals: expected "<span><a>test</a><frame><b></b></span>" but got "<span><a>test</a><frame></frame><b></b></span>"
FAIL outerHTML void as second child with following siblings hr assert_equals: expected "<span><a>test</a><hr><b></b></span>" but got "<span><a>test</a><hr></hr><b></b></span>"
FAIL outerHTML void as second child with following siblings img assert_equals: expected "<span><a>test</a><img><b></b></span>" but got "<span><a>test</a><img></img><b></b></span>"
FAIL outerHTML void as second child with following siblings input assert_equals: expected "<span><a>test</a><input><b></b></span>" but got "<span><a>test</a><input></input><b></b></span>"
FAIL outerHTML void as second child with following siblings keygen assert_equals: expected "<span><a>test</a><keygen><b></b></span>" but got "<span><a>test</a><keygen></keygen><b></b></span>"
FAIL outerHTML void as second child with following siblings link assert_equals: expected "<span><a>test</a><link><b></b></span>" but got "<span><a>test</a><link></link><b></b></span>"
FAIL outerHTML void as second child with following siblings menuitem assert_equals: expected "<span><a>test</a><menuitem><b></b></span>" but got "<span><a>test</a><menuitem><a><!--abc--></a><b><c>abc</c></b></menuitem><b></b></span>"
FAIL outerHTML void as second child with following siblings meta assert_equals: expected "<span><a>test</a><meta><b></b></span>" but got "<span><a>test</a><meta></meta><b></b></span>"
FAIL outerHTML void as second child with following siblings param assert_equals: expected "<span><a>test</a><param><b></b></span>" but got "<span><a>test</a><param></param><b></b></span>"
FAIL outerHTML void as second child with following siblings source assert_equals: expected "<span><a>test</a><source><b></b></span>" but got "<span><a>test</a><source></source><b></b></span>"
FAIL outerHTML void as second child with following siblings track assert_equals: expected "<span><a>test</a><track><b></b></span>" but got "<span><a>test</a><track></track><b></b></span>"
FAIL outerHTML void as second child with following siblings wbr assert_equals: expected "<span><a>test</a><wbr><b></b></span>" but got "<span><a>test</a><wbr></wbr><b></b></span>"
FAIL outerHTML void as last child with preceding siblings area assert_equals: expected "<span><a>test</a><b></b><area></span>" but got "<span><a>test</a><b></b><area></area></span>"
FAIL outerHTML void as last child with preceding siblings base assert_equals: expected "<span><a>test</a><b></b><base></span>" but got "<span><a>test</a><b></b><base></base></span>"
FAIL outerHTML void as last child with preceding siblings basefont assert_equals: expected "<span><a>test</a><b></b><basefont></span>" but got "<span><a>test</a><b></b><basefont></basefont></span>"
FAIL outerHTML void as last child with preceding siblings bgsound assert_equals: expected "<span><a>test</a><b></b><bgsound></span>" but got "<span><a>test</a><b></b><bgsound></bgsound></span>"
FAIL outerHTML void as last child with preceding siblings br assert_equals: expected "<span><a>test</a><b></b><br></span>" but got "<span><a>test</a><b></b><br></br></span>"
FAIL outerHTML void as last child with preceding siblings col assert_equals: expected "<span><a>test</a><b></b><col></span>" but got "<span><a>test</a><b></b><col></col></span>"
FAIL outerHTML void as last child with preceding siblings embed assert_equals: expected "<span><a>test</a><b></b><embed></span>" but got "<span><a>test</a><b></b><embed></embed></span>"
FAIL outerHTML void as last child with preceding siblings frame assert_equals: expected "<span><a>test</a><b></b><frame></span>" but got "<span><a>test</a><b></b><frame></frame></span>"
FAIL outerHTML void as last child with preceding siblings hr assert_equals: expected "<span><a>test</a><b></b><hr></span>" but got "<span><a>test</a><b></b><hr></hr></span>"
FAIL outerHTML void as last child with preceding siblings img assert_equals: expected "<span><a>test</a><b></b><img></span>" but got "<span><a>test</a><b></b><img></img></span>"
FAIL outerHTML void as last child with preceding siblings input assert_equals: expected "<span><a>test</a><b></b><input></span>" but got "<span><a>test</a><b></b><input></input></span>"
FAIL outerHTML void as last child with preceding siblings keygen assert_equals: expected "<span><a>test</a><b></b><keygen></span>" but got "<span><a>test</a><b></b><keygen></keygen></span>"
FAIL outerHTML void as last child with preceding siblings link assert_equals: expected "<span><a>test</a><b></b><link></span>" but got "<span><a>test</a><b></b><link></link></span>"
FAIL outerHTML void as last child with preceding siblings menuitem assert_equals: expected "<span><a>test</a><b></b><menuitem></span>" but got "<span><a>test</a><b></b><menuitem><a><!--abc--></a><b><c>abc</c></b></menuitem></span>"
FAIL outerHTML void as last child with preceding siblings meta assert_equals: expected "<span><a>test</a><b></b><meta></span>" but got "<span><a>test</a><b></b><meta></meta></span>"
FAIL outerHTML void as last child with preceding siblings param assert_equals: expected "<span><a>test</a><b></b><param></span>" but got "<span><a>test</a><b></b><param></param></span>"
FAIL outerHTML void as last child with preceding siblings source assert_equals: expected "<span><a>test</a><b></b><source></span>" but got "<span><a>test</a><b></b><source></source></span>"
FAIL outerHTML void as last child with preceding siblings track assert_equals: expected "<span><a>test</a><b></b><track></span>" but got "<span><a>test</a><b></b><track></track></span>"
FAIL outerHTML void as last child with preceding siblings wbr assert_equals: expected "<span><a>test</a><b></b><wbr></span>" but got "<span><a>test</a><b></b><wbr></wbr></span>"
